Refactoring: Added EntityFactory and moved parsing
