+++ custom_components/enpal_webparser/utils.py
@@ -18,6 +18,12 @@
 
 import re
 from typing import Optional, Tuple, Dict
+from bs4 import BeautifulSoup, Tag
+from datetime import datetime
+from typing import List, Dict, Any, Optional
+
+from .const import (
+    UNIT_DEVICE_CLASS_MAP, ENPAL_TIMESTAMP_FORMAT,DEFAULT_UNITS)
 
 def make_id(name: str) -> str:
     """Generate a Home Assistant-friendly unique id from a name."""
@@ -88,3 +94,54 @@ def normalize_value_and_unit(
         unit_out = default_units.get(device_class)
 
     return value_clean, unit_out
+
+
+def parse_enpal_html_sensors(
+    html: str, 
+    groups: list[str]
+) -> List[Dict[str, Any]]:
+    """Parst Enpal HTML und liefert Sensor-Data-Listen für Home Assistant."""
+    soup = BeautifulSoup(html, 'html.parser')
+    cards = soup.find_all("div", class_="card")
+    sensors = []
+    for card in cards:
+        if not isinstance(card, Tag):
+            continue
+        h2_tag = card.find("h2")
+        if h2_tag is not None:
+            group = h2_tag.text.strip()
+        else:
+            continue
+        if group not in groups:
+            continue
+
+        rows = card.find_all("tr")[1:]
+        for row in rows:
+            if not isinstance(row, Tag):
+                continue
+            cols = row.find_all("td")
+            if len(cols) >= 2:
+                raw_name = cols[0].text.strip()
+                value_raw = cols[1].text.strip()
+                unit, device_class = get_class_and_unit(value_raw, UNIT_DEVICE_CLASS_MAP)
+                value_clean, unit = normalize_value_and_unit(value_raw, unit, device_class, DEFAULT_UNITS)
+
+                timestamp_str = cols[2].text.strip() if len(cols) > 2 else None
+                timestamp_iso = None
+                if timestamp_str:
+                    try:
+                        dt = datetime.strptime(timestamp_str, ENPAL_TIMESTAMP_FORMAT)
+                        timestamp_iso = dt.isoformat()
+                    except ValueError:
+                        timestamp_iso = timestamp_str
+
+                sensor_data = {
+                    "name": friendly_name(group, raw_name),
+                    "value": value_clean,
+                    "unit": unit,
+                    "device_class": device_class,
+                    "enabled": group in groups,
+                    "enpal_last_update": timestamp_iso,
+                }
+                sensors.append(sensor_data)
+    return sensors
